+++ hyojaenft-front/src/App.tsx
@@ -1,0 +1,12 @@
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header"></header>
+      <body>
+        <h1>jaewi Hi</h1>
+      </body>
+    </div>
+  );
+}
+
+export default App;
